+++ sw.js
@@ -1,10 +1,11 @@
-const CACHE = 'n5-v105';
+const CACHE = 'n5-v106';
 const ASSETS = [
   './',
   './index.html',
   './style.css',
   './app.js',
   './conjugate.js',
+  './furigana.js',
   './cloze.js',
   './pos.js',
   './srs.js',